* origin/release/v5.0: (1207 commits)
  feat(spi_lcd): control DC level for cmd, param, color phase separately
  Fix(esp_wifi): Add the missing header file to the wifi header files
  ci: add sbom validation using pre-commit's post-commit stage
  change(uart): improved the internal logic of uart_read_bytes
  fix(ble_mesh): reject node-reseting when provision link is active (V5.0)
  mbedtls: define MBEDTLS_SSL_CID_TLS1_3_PAD_GRANULARITY for CID padding
  fix(bt/bluedroid): Update ble gattc gatts coex example
  fix(bt/bluedroid): Optimize ble remove all bonded devices function
  fix(bt/bluedroid): Update ble ibeacon example document
  fix(bt/bluedroid): Fix BLE robost cashing config
  fix(esp_wifi): Fix timer index out of bound issue causing execution of timer function to fail
  fix(bt/bluedroid): Fix ble app unregister no event
  fix(nimble): Handled race condition for conn variables being set to NULL(v5.0)
  fix(phy): Fix ble tx ifs time for ESP32
  fix(examples): Fix encoding issue in esp_local_ctrl's proto script
  fix(bt/bluedroid): Disabled Sniff Subrating temporarily as it did not work well
  fix(bt/bluedroid): Changed the default air mode to invalid value
  fix(bt/bluedroid): Removed redundant operation of SCO clean up
  fix(bt/bluedroid): Corrected the definitions of HF Client callback events
  fix(bt/bluedroid): Avoided crash of LoadProhibited during HFP AG deinitialization
  ...

# Conflicts:
#	tools/idf.py

+++ components/esp_hw_support/include/esp_cpu.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2020-2022 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2020-2023 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -19,6 +19,7 @@
 #endif
 #include "esp_intr_alloc.h"
 #include "esp_err.h"
+#include "esp_attr.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -413,9 +414,9 @@ FORCE_INLINE_ATTR void esp_cpu_intr_edge_ack(int intr_num)
 {
     assert(intr_num >= 0 && intr_num < SOC_CPU_INTR_NUM);
 #ifdef __XTENSA__
-    xthal_set_intclear(1 << intr_num);
+    xthal_set_intclear((unsigned) (1 << intr_num));
 #else
-    rv_utils_intr_edge_ack(intr_num);
+    rv_utils_intr_edge_ack((unsigned) intr_num);
 #endif
 }
 
@@ -467,9 +468,15 @@ esp_err_t esp_cpu_clear_breakpoint(int bp_num);
  * the CPU accesses (according to the trigger type) on a certain memory range.
  *
  * @note Overwrites previously set watchpoint with same watchpoint number.
+ *       On RISC-V chips, this API uses method0(Exact matching) and method1(NAPOT matching) according to the
+ *       riscv-debug-spec-0.13 specification for address matching.
+ *       If the watch region size is 1byte, it uses exact matching (method 0).
+ *       If the watch region size is larger than 1byte, it uses NAPOT matching (method 1). This mode requires
+ *       the watching region start address to be aligned to the watching region size.
+ *
  * @param wp_num Hardware watchpoint number [0..SOC_CPU_WATCHPOINTS_NUM - 1]
- * @param wp_addr Watchpoint's base address
- * @param size Size of the region to watch. Must be one of 2^n, with n in [0..6].
+ * @param wp_addr Watchpoint's base address, must be naturally aligned to the size of the region
+ * @param size Size of the region to watch. Must be one of 2^n and in the range of [1 ... SOC_CPU_WATCHPOINT_MAX_REGION_SIZE]
  * @param trigger Trigger type
  * @return ESP_ERR_INVALID_ARG on invalid arg, ESP_OK otherwise
  */
